split42 reconstruct 1/7: rename folder sources split72.* -> split42.*
- git mv split72.c -> split42.c, split72.h -> split42.h (QMK auto-includes the
  <keyboard>.c/.h named after the folder, so these MUST be split42.*).
- fix split42.c's include split72.h -> split42.h.
- drop split72-only extras: vphya.json (split72 VIA/Vial layout) + keymaps/revision2.
Content of the two renamed files is still split72's; later steps adapt it.
