--- keyboards/polykybd/split42/keymaps/revision2/config.h
@@ -1,4 +1,0 @@
-#define ENABLE_COMPILE_KEYCODE
-
-#define EECONFIG_USER_DATA_SIZE 65   // +1 for poly_eeconf_t.glyph_script (within POLY_EECONFIG_USER_RESERVED=128, no keymap relocation)
-
