--- keyboards/polykybd/split42/keymaps/revision2/keymap.c
@@ -1,1 +1,0 @@
-#include "../default/keymap.c"
